Expose SIO type
